+++ internal/service/store/service.go
@@ -12,6 +12,7 @@ import (
 	"github.com/dv-net/dv-merchant/internal/service/exrate"
 	"github.com/dv-net/dv-merchant/internal/service/notify"
 	"github.com/dv-net/dv-merchant/internal/service/processing"
+	"github.com/dv-net/dv-merchant/internal/service/setting"
 	"github.com/dv-net/dv-merchant/internal/service/transactions"
 	"github.com/dv-net/dv-merchant/internal/service/wallet"
 	"github.com/dv-net/dv-merchant/internal/service/webhook"
@@ -54,6 +55,7 @@ type Service struct {
 	wallets             wallet.IWalletService
 	notificationService notify.INotificationService
 	processingSvc       processing.IProcessingOwner
+	settingSvc          setting.ISettingService
 }
 
 var _ IStore = (*Service)(nil)
@@ -72,6 +74,7 @@ func New(
 	rateLimit rate.Limiter,
 	rateLimitEnabled bool,
 	processingSvc processing.IProcessingOwner,
+	settingSvc setting.ISettingService,
 ) *Service {
 	srv := &Service{
 		storage:             storage,
@@ -85,6 +88,7 @@ func New(
 		rateLimitEnabled:    rateLimitEnabled,
 		notificationService: notificationService,
 		processingSvc:       processingSvc,
+		settingSvc:          settingSvc,
 	}
 	// register event
 	srv.eventListener.Register(transactions.DepositReceivedEventType, srv.handleDepositReceived)
@@ -179,6 +183,11 @@ func (s *Service) CreateStore(ctx context.Context, dto CreateStore, user *models
 		}
 		// create store secret
 		_, err = s.GenerateSecret(ctx, st.ID, repos.WithTx(tx))
+		if err != nil {
+			return err
+		}
+		// create store settings
+		err = s.CreateStoreSettings(ctx, st.ID, repos.WithTx(tx))
 		if err != nil {
 			return err
 		}